test(instrumentation): move test common code to new base class
